#6: Solving some problems with mole, ant and the destruction of the towers.

+++ src/main/java/edu/fiuba/algo3/TypeData/Credits/Credits.java
@@ -1,9 +1,9 @@
 package edu.fiuba.algo3.TypeData.Credits;
 
 public class Credits {
-    int credits;
+    double credits;
 
-    public Credits(int credits) {
+    public Credits(double credits) {
         this.credits = credits;
     }
